bypassset: пакет вместо selective, набор AWGM-BYPASS
