+++ internal/singbox/router/bypassset/deps_check_test.go
@@ -1,4 +1,4 @@
-package selective
+package bypassset
 
 import (
 	"errors"
