+++ internal/singbox/router/bypassset/ipset_staging.go
@@ -1,4 +1,4 @@
-package selective
+package bypassset
 
 import (
 	"bytes"
@@ -13,7 +13,7 @@ import (
 
 // StagingSetName is the scratch ipset populated during rebuild and atomically
 // swapped with SetName when the pipeline completes.
-const StagingSetName = "AWGM-SELECTIVE-STG"
+const StagingSetName = "AWGM-BYPASS-STG"
 
 // IpsetChunkSize is how many entries are batched into one ipset restore call.
 const IpsetChunkSize = 512
@@ -21,8 +21,7 @@ const IpsetChunkSize = 512
 // ipsetRestoreTimeout ограничивает один вызов `ipset restore` на чанк из
 // IpsetChunkSize записей. Медленный роутер под нагрузкой (256–512MB MIPS)
 // легально укладывает чанк не за секунды, а за минуты; чанк конечен (512
-// записей), поэтому зависание команды ловит этот exec-таймаут, а не stall
-// guard пересборки (rebuildStallTimeout подобран не меньше этого шага).
+// записей), поэтому щедрый потолок безопасен.
 const ipsetRestoreTimeout = 180 * time.Second
 
 var ipsetChunkPool = sync.Pool{
@@ -66,11 +65,6 @@ func ChunkedAddStaging(ctx context.Context, cidrs []string) error {
 	return chunkedAddToSet(ctx, StagingSetName, cidrs)
 }
 
-// ChunkedAddLive appends entries to the live set (CDN refresh path).
-func ChunkedAddLive(ctx context.Context, cidrs []string) error {
-	return chunkedAddToSet(ctx, SetName, cidrs)
-}
-
 func chunkedAddToSet(ctx context.Context, setName string, cidrs []string) error {
 	if len(cidrs) == 0 {
 		return nil
@@ -105,14 +99,8 @@ func addEntriesToSet(ctx context.Context, setName string, cidrs []string) error 
 	if writeRestoreLines(b, setName, cidrs) == 0 {
 		return nil
 	}
-	// Прогресс stall guard'у до и после restore-команды (ProgressTouch —
-	// no-op вне пересборки): «начали операцию» — тоже прогресс, зависание
-	// самой команды ловит её exec-таймаут ipsetRestoreTimeout.
-	touch := ProgressTouch(ctx)
-	touch()
 	res, err := sysexec.RunWithOptions(ctx, bin, []string{"restore", "-exist"},
 		sysexec.Options{Stdin: b, Timeout: ipsetRestoreTimeout})
-	touch()
 	if err != nil {
 		return sysexec.FormatError(res, fmt.Errorf("ipset restore: %w", err))
 	}
@@ -126,7 +114,7 @@ func addEntriesToSet(ctx context.Context, setName string, cidrs []string) error 
 func writeRestoreLines(b *bytes.Buffer, setName string, cidrs []string) int {
 	valid := 0
 	for _, raw := range cidrs {
-		entry := normalizeEntry(raw)
+		entry := NormalizeEntry(raw)
 		if entry == "" {
 			continue
 		}
